rename panelapi to remote
